Use Float for confidence in annotations
This matches the type used in the database schema

+++ app-backend/datamodel/src/main/scala/com/azavea/rf/datamodel/Annotation.scala
@@ -23,7 +23,7 @@ case class Annotation(
   label: String,
   description: Option[String],
   machineGenerated: Option[Boolean],
-  confidence: Option[Double],
+  confidence: Option[Float],
   quality: Option[AnnotationQuality],
   geometry: Option[Projected[Geometry]]
 ) extends GeoJSONSerializable[Annotation.GeoJSON] {
@@ -62,7 +62,7 @@ case class AnnotationProperties(
     label: String,
     description: Option[String],
     machineGenerated: Option[Boolean],
-    confidence: Option[Double],
+    confidence: Option[Float],
     quality: Option[AnnotationQuality]
 )
 
@@ -72,7 +72,7 @@ case class AnnotationPropertiesCreate(
     label: String,
     description: Option[String],
     machineGenerated: Option[Boolean],
-    confidence: Option[Double],
+    confidence: Option[Float],
     quality: Option[AnnotationQuality]
 )
 
@@ -104,7 +104,7 @@ object Annotation {
         label: String,
         description: Option[String],
         machineGenerated: Option[Boolean],
-        confidence: Option[Double],
+        confidence: Option[Float],
         quality: Option[AnnotationQuality],
         geometry: Option[Projected[Geometry]]
     ) extends OwnerCheck {
